extracted VehicleState
